lucky_api/pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "Aadarsha",email = "aadarshakarki000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "web3 (>=7.15.0,<8.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-treebeard (>=5.0.5,<6.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "django-celery-beat (>=2.9.0,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
